moved endorse button to endorsements folder
