+++ client/src/app/components/endorsements/endorse-assertion-button/endorse-assertion-button.component.ts
